requirements.txt
agent-framework>=1.1.0
agent-framework-foundry-hosting>=1.0.0a260420
python-dotenv
azure-identity

# Azure Monitor / OpenTelemetry
azure-monitor-opentelemetry-exporter>=1.0.0b46
opentelemetry-sdk>=1.39.0
opentelemetry-api>=1.39.0